somedemo

+++ Python-study/pyday01_03ZY/2.py
@@ -1,7 +1,8 @@
+#coding=UTF-8
 #2、输入长方形的长和宽，计算长方形的面积和周长
 L=float(input("输入长方形的长:"))
 W=float(input("输入长方形的宽:"))
-S=L*W
-D=(L+W)*2
+S=int(L*W)
+D=int((L+W)*2)
 print(S)
 print(D)
